MondayResponse, Code Refactoring + Updates
MondayResponse<T> has been added to each method,
T will be the return type. By having this, we will be able to handle errors a lot easier as MondayResponse contains both IsSucessful and an errors collection.

GetItemsQueryBuilder & UnsupportedTypes  moved to Utililies.

Item record Id is now a ulong instead of a string as it made more sense and shouldnt be a problem as only column  values id can be a non numeric value...I think.

DeleteItemsAsync has been added