Email Column Added
Email Column Has Been Added As Per #2
Users have the ability to enter the Id, email and message if any.

Currently not checking if its a valid email address or not (will add at a later date), please use with caution.

Filtering items by email address to be added when items-query is implemented.